Allow part volume to be higher than 100%

+++ src/pt2_synth.h
@@ -146,7 +146,7 @@ typedef struct program_t
 typedef struct part_t
 {
     uint8_t program;
-    uint8_t volume;
+    uint16_t volume;
     uint16_t sampleRate;
     uint16_t offset;
 } part_t;
